+++ pyrevolut/api/foreign_exchange/post/exchange_money.py
@@ -1,6 +1,5 @@
 from typing import Annotated
 from uuid import UUID
-from decimal import Decimal
 
 from pydantic import BaseModel, Field, model_validator, ConfigDict
 from pydantic_extra_types.currency_code import Currency
@@ -47,7 +46,7 @@ class ExchangeMoney:
                 Field(description="The currency to sell in ISO 4217 format."),
             ]
             amount: Annotated[
-                Decimal | None,
+                float | None,
                 Field(
                     description="The amount of currency. Specify ONLY if you want to sell currency."
                 ),
@@ -67,7 +66,7 @@ class ExchangeMoney:
                 Field(description="The currency to buy in ISO 4217 format."),
             ]
             amount: Annotated[
-                Decimal | None,
+                float | None,
                 Field(
                     description="The amount of currency. Specify ONLY if you want to buy currency."
                 ),
